rename views folder to pages, make it more understandable
