+++ src/router/index.js
@@ -1,6 +1,6 @@
 import Vue from 'vue';
 import VueRouter from 'vue-router';
-import Home from '../views/Home.vue';
+import Home from '../pages/Home.vue';
 
 Vue.use(VueRouter);
 
@@ -16,7 +16,7 @@ const routes = [
     // route level code-splitting
     // this generates a separate chunk (detail.[hash].js) for this route
     // which is lazy-loaded when the route is visited.
-    component: () => import('../views/NewsDetail.vue'),
+    component: () => import('../pages/NewsDetail.vue'),
   },
 ];
